+++ crates/yast-core/src/lib.rs
@@ -1,3 +1,5 @@
+//! `yast-core` is YASX's main lib.
+
 pub mod defaults;
 pub mod layout;
 pub mod lua;